+++ doc/source/conf.py
@@ -18,12 +18,12 @@ sys.path.insert(0, os.path.abspath("../"))
 
 # -- Project information -----------------------------------------------------
 
-project = "Recpack"
-copyright = "2021, Lien Michiels"
-author = "Lien Michiels"
+project = "RecPack"
+copyright = "2022, Lien Michiels"
+author = "Lien Michiels and Robin Verachtert"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.1"
+release = "0.3.0"
 
 
 # -- General configuration ---------------------------------------------------
@@ -58,6 +58,10 @@ html_theme = "sphinx_rtd_theme"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+
+html_css_files = [
+    'css/custom.css',
+]
 
 # -- Options for MathJax --------
 mathjax_config = {