+++ recpack/scenarios/__init__.py
@@ -17,9 +17,9 @@ according to one, simple criterion, e.g. "Fold in is all interactions before T,
 hold out all interactions after T".
 
 Should you want to implement a new scenario that is not yet supported,
-you can use the splitter submodule to help you build it from the ground up.
+you can use the splitters submodule to help you build it from the ground up.
 
-.. currentmodule:: recpack.splitters
+.. currentmodule:: recpack.scenarios
 
 .. contents:: Table of Contents
     :depth: 2