+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * __pipelines__
     * PipelineBuilder's `set_train_data` function has been replaced with `set_full_training_data` and `set_validation_training_data` to set the two separate training datasets.
     * Added `set_data_from_scenario` function to `PipelineBuilder`, which makes setting data easy based on a split scenario.
+    * Updated `get_metrics` function, which returns the metrics as a dataframe, rather than a dict.
+
+* Several module changes were made:
+    * splitters module was restructured. Base module is called scenarios. Inside scenarios the splitters submodule contains the splitters functionality.
+    * data module was removed, and the submodules were turned into modules.  (`dataset` and `matrix`)
+
 
 * Several module changes were made:
     * splitters module was restructured. Base module is called scenarios. Inside scenarios the splitters submodule contains the splitters functionality.